editcontract template and "tariff options edit" improved
editcontract - just jsp without logic
"tariff options edit" : dependent and required options are added and
removed automatically

+++ eCareIS/src/main/java/service/TariffServiceImpl.java
@@ -4,6 +4,8 @@ import DAO.*;
 import entity.Contract;
 import entity.Option;
 import entity.Tariff;
+import org.apache.log4j.Logger;
+import service.DTO.OptionDTO;
 import service.DTO.TariffDTO;
 import utils.Constants;
 import utils.EntityManagerFactorySingleton;
@@ -16,9 +18,12 @@ import java.util.Set;
  * Created by Alexey on 04.07.2015.
  */
 public class TariffServiceImpl implements TariffService{
+
     TariffDAO tariffDAO = new TariffDAOImpl(EntityManagerFactorySingleton.getInstance());
     OptionDAO optionDAO = new OptionDAOImpl(EntityManagerFactorySingleton.getInstance());
     ContractDAO contractDAO = new ContractDAOImpl(EntityManagerFactorySingleton.getInstance());
+    OptionService optionService = new OptionServiceImpl();
+    Logger logger = Logger.getLogger(TariffServiceImpl.class);
 
     @Override
     public TariffDTO getTariffById(Integer tariffId) {
@@ -42,23 +47,33 @@ public class TariffServiceImpl implements TariffService{
 
     @Override
     public void addOptionAsPossibleForTariff(Integer tariffId, Integer optionId) {
-        Option option = optionDAO.get(optionId);
+        Set<OptionDTO> allRequiredOptionTree = optionService.getRequiredOptionTree(optionId);
         Tariff tariff = tariffDAO.get(tariffId);
-        tariff.getPossibleOption().add(option);
+        for (OptionDTO requiredOption : allRequiredOptionTree){
+            tariff.getPossibleOption().add(optionDAO.get(requiredOption.getOptionId()));
+        }
         tariffDAO.update(tariff);
     }
 
     @Override
-    public void removeOptionAsPossibleForTariff(Integer tariffId, Integer optionId) {
-        Option option = optionDAO.get(optionId);
+    public void removeOptionAndAllDependentOptionsTreeAsPossibleForTariff(
+            Integer tariffId, Integer optionId) {
+
+        Set<OptionDTO> allDependentOptionTree = optionService.getDependentOptionTree(optionId);
         Tariff tariff = tariffDAO.get(tariffId);
-        tariff.getPossibleOption().remove(option);
+        for (OptionDTO dependentOption : allDependentOptionTree){
+            tariff.getPossibleOption().remove(optionDAO.get(dependentOption.getOptionId()));
+        }
         tariffDAO.update(tariff);
     }
 
     @Override
     public void removeTariffAndMoveContractsToBaseTariff(Integer tariffId) {
-        //todo make a check it is now a base tariff!
+//        TODO: delete all options connected to the contract
+        if (tariffId==Constants.DEFAULT_TARIFF_ID){
+            logger.warn("Attempt to delete default tariff");
+            return;
+        }
         Tariff tariff = tariffDAO.get(tariffId);
         Set<Contract> contractsHasThisTariff = tariff.getContractHasThisTariff();
         tariffDAO.delete(tariffId);
@@ -67,4 +82,5 @@ public class TariffServiceImpl implements TariffService{
             contractDAO.update(contract);
         }
     }
+
 }
